// actyondatafront/src/app/shared/models/titulos/titulo-modal.ts
export class TituloModel {
    TITULO_ID: number;
    TIPO_TITULO_ID: number;
    DEVEDOR_ID: number;
    NUMERO_CONTRATO: string;
    NUMERO_PRESTACAO: number;
    QTDE_PRESTACAO: number;
    NUMERO_DOCUMENTO: string;
    VALOR: number;
    DATA_VENCIMENTO: Date;
    PRODUTO_ID: string;
    OBSE: string;
    CONT_ID: number;
    NOME_ALUNO: string;
    DESCRICAO: string;
    DATA_PAGTO: Date;
    TIPO_BAIXA: number;
    VALOR_ADICIONAL1: number;
    VALOR_ADICIONAL2: number;
}